[level 1] Title: 부족한 금액 계산하기, Time: 0.10 ms, Memory: 84.3 MB -BaekjoonHub

+++ 프로그래머스/1/82612. 부족한 금액 계산하기/부족한 금액 계산하기.java
@@ -3,7 +3,7 @@ class Solution {
         
         long total = 0; // 놀이기구 이용 금액
         for(int i=1; i<=count; i++) {
-            total += (long) i * price;
+            total += i * price;
         }
         
         long result = total - money;
